Crop parametrization information moved to separate cards.
