+++ RUFAS/routines/crop/crop.py
@@ -42,8 +42,9 @@ This module needs the following inputs in order to operate correctly:
 ################################################################################
 
 from math import acos, asin, sin, tan, pi
+from .crop_types import alfalfa, corn, soybean
 from . import heat_units, leaf_area_index, root_development, biomass, yields, \
-    phosphorus_uptake, nitrogen_uptake, transpiration, growth_constraints
+    phosphorus_uptake, nitrogen_uptake, growth_constraints
 
 
 # -------------------------------------------------------------------------------
@@ -178,9 +179,9 @@ def is_kill_year(crop, time):
 class Crop:
     def __init__(self, data, time):
         self.init_crop = InitCrop(data)
-        self.alfalfa = Alfalfa(data)
-        self.corn = Corn(data)
-        self.soy = Soybean(data)
+        self.alfalfa = alfalfa.Alfalfa(data)
+        self.corn = corn.Corn(data)
+        self.soy = soybean.Soybean(data)
 
         # Dormancy for perennial crops
         self.latitude = abs(data['latitude'])
@@ -191,7 +192,7 @@ class Crop:
         self.crops_list = [self.alfalfa, self.corn, self.soy]
         self.current_crop = self.init_crop
 
-        self.grow_regimen = [self.init_crop for x in range(0, len(time.years))]
+        self.grow_regimen = [self.init_crop for _ in range(0, len(time.years))]
 
         # Each crop in crops_list has a list of grow years. This loop iterates
         # through those lists and populates a grow regimen with the crop grown
@@ -369,489 +370,6 @@ class InitCrop:
         self.HI_opt = 0
 
         self.harvest_eff = 0
-
-        self.gamma_wu = 0
-
-        self.bio_AG = 0
-        self.yield_max = 0
-        self.yield_actual = 0
-        self.yield_N = 0
-        self.yield_P = 0
-
-        self.yield_annual = 0
-
-
-#
-# Crop object populated with Corn data
-#
-class Corn:
-
-    def __init__(self, data):
-        """GENERAL PLANT INFO"""
-
-        corn_data = data['corn']
-        self.grow_years = corn_data['grow_years']
-        self.repeat = corn_data['repeat']
-        self.planting_date = corn_data['planting_date']
-        self.harvest_date = corn_data['harvest_date']
-        self.harvest_type = corn_data['harvest_type']
-
-        self.crop_name = 'corn'
-        self.crop_type = 'annual'
-
-        self.kill_day = -1
-        self.kill_year = True
-        self.planted = False
-        self.growing = False
-
-        self.fix_nitrogen = False
-
-        # ===================================================================
-        ''' HEAT UNIT DATA '''
-
-        # Inputs
-        self.T_base_min = 10
-        self.T_base_max = 30
-        self.PHU = 1200
-
-        # Internally calculated inputs
-        self.accumulated_HU = 0.0
-        self.prev_accumulated_HU = 0.0
-
-        # Outputs
-        self.fr_PHU = 0.0
-        self.prev_fr_PHU = 0.0
-
-        # ===================================================================
-        ''' LEAF AREA INDEX (LAI) DATA '''
-
-        # Inputs
-        self.fr_PHU_1 = 0.15
-        self.fr_PHU_2 = 0.50
-        self.fr_LAI_1 = 0.05
-        self.fr_LAI_2 = 0.95
-        self.fr_PHU_sen = 0.90
-        self.fr_PHU_harvest = 1.2
-        self.LAI_max = 3
-        self.LAI_min = 0
-
-        # Internally calculated inputs
-        self.prev_fr_LAI_max = 0
-        self.fr_LAI_max = 0
-
-        # Outputs
-        self.prev_LAI_actual = 0
-        self.LAI_actual = 0
-
-        # ===================================================================
-        ''' ROOT DEPTH DATA '''
-
-        # Inputs
-        self.z_root_max = 2000  # maximum depth of root development
-
-        # Internally calculated inputs
-        self.fr_root = 0
-
-        # Outputs
-        self.z_root = 0
-
-        # ===================================================================
-        ''' BIOMASS DATA '''
-
-        # Inputs
-        self.kl = 0.65
-        self.RUE = 39
-        self.T_opt = 25
-
-        # Internally calculated inputs
-        self.gamma_reg = 0
-        self.d_biomass_max = 0
-        self.d_biomass_actual = 0.0
-
-        # Outputs
-        self.biomass_actual = 0
-        self.prev_biomass_actual = 0
-
-        # ===================================================================
-        ''' Soil Water Uptake Data '''
-
-        self.beta_w = 10  # water-use distribution parameter
-        self.epco = 0.5
-
-        self.water_actual_up = 0
-        self.water_uptake_each_layer = []
-
-        # ===================================================================
-        ''' Nitrogen Uptake Data '''
-
-        self.beta_n = 10
-
-        self.bio_N_opt = 0
-        self.bio_N = 0
-
-        self.fr_n1 = 0.047
-        self.fr_n2 = 0.0177
-        self.fr_n3 = 0.0138
-        self.fr_n3ish = 0.01381
-
-        self.fr_N = 0
-        self.fr_N_up = 0
-        self.N_up = 0
-        self.act_N_up_each_layer = []
-        self.N_actual_up = 0
-
-        # ===================================================================
-        ''' Phosphorus Uptake Data '''
-
-        self.beta_p = 10
-
-        self.bio_P_opt = 0
-        self.bio_P = 0
-
-        self.fr_PHU_50 = 0.5
-        self.fr_PHU_100 = 1.0
-        self.fr_p1 = 0.0048
-        self.fr_p2 = 0.0018
-        self.fr_p3 = 0.0014
-        self.fr_p3ish = 0.00141
-
-        self.fr_P = 0
-        self.P_up = 0
-        self.act_P_up_each_layer = []
-        self.P_act_up = 0
-
-        # ===================================================================
-        ''' Yields Data '''
-
-        self.HI_max = 0
-        self.HI_min = 0.3
-        self.HI_actual = 0
-        self.HI_opt = 0.6
-
-        self.harvest_eff = 0.9
-
-        self.gamma_wu = 0
-
-        self.bio_AG = 0
-        self.yield_max = 0
-        self.yield_actual = 0
-        self.yield_N = 0
-        self.yield_P = 0
-
-        self.yield_annual = 0
-
-
-#
-# Crop object populated with Soybean data
-#
-class Soybean:
-
-    def __init__(self, data):
-        """GENERAL PLANT INFO"""
-
-        soy_data = data['soybean']
-        self.grow_years = soy_data['grow_years']
-        self.repeat = soy_data['repeat']
-        self.planting_date = soy_data['planting_date']
-        self.harvest_date = soy_data['harvest_date']
-        self.harvest_type = soy_data['harvest_type']
-
-        self.crop_name = 'soybean'
-        self.crop_type = 'annual'
-
-        self.kill_day = -1
-        self.kill_year = True
-        self.planted = False
-        self.growing = False
-
-        self.fix_nitrogen = True
-        # ===================================================================
-        ''' HEAT UNIT DATA '''
-
-        # Inputs
-        self.T_base_min = 10
-        self.T_base_max = 30  # corn
-        self.PHU = 1150
-
-        # Internally calculated inputs
-        self.accumulated_HU = 0.0
-        self.prev_accumulated_HU = 0.0
-
-        # Outputs
-        self.fr_PHU = 0.0
-        self.prev_fr_PHU = 0.0
-
-        # ===================================================================
-        ''' LEAF AREA INDEX (LAI) DATA '''
-
-        # Inputs
-        self.fr_PHU_1 = 0.15
-        self.fr_PHU_2 = 0.50
-        self.fr_LAI_1 = 0.05
-        self.fr_LAI_2 = 0.95
-        self.fr_PHU_sen = 0.9
-        self.fr_PHU_harvest = 1.0  # TODO: If soybean has drydown, this is 1.2
-        self.LAI_max = 3
-        self.LAI_min = 0
-
-        # Internally calculated inputs
-        self.prev_fr_LAI_max = 0
-        self.fr_LAI_max = 0
-
-        # Outputs
-        self.prev_LAI_actual = 0
-        self.LAI_actual = 0
-
-        # ===================================================================
-        ''' ROOT DEPTH DATA '''
-
-        # Inputs
-        self.z_root_max = 1700  # maximum depth of root development
-
-        # Internally calculated inputs
-        self.fr_root = 0
-
-        # Outputs
-        self.z_root = 0
-
-        # ===================================================================
-        ''' BIOMASS DATA '''
-
-        # Inputs
-        self.kl = 0.45
-        self.RUE = 25
-        self.T_opt = 25
-
-        # Internally calculated inputs
-        self.gamma_reg = 0
-        self.d_biomass_max = 0
-        self.d_biomass_actual = 0.0
-
-        # Outputs
-        self.biomass_actual = 0
-        self.prev_biomass_actual = 0
-
-        # ===================================================================
-        ''' Soil Water Uptake Data '''
-
-        self.beta_w = 10  # water-use distribution parameter  # corn
-        self.epco = 0.5  # corn
-
-        # Outputs
-        self.prev_LAI_actual = 0
-        self.LAI_actual = 0
-
-        # ===================================================================
-        ''' Nitrogen Uptake Data '''
-
-        self.beta_n = 10  # corn
-
-        self.bio_N_opt = 0
-        self.bio_N = 0
-
-        self.fr_n1 = 0.0524
-        self.fr_n2 = 0.0265
-        self.fr_n3 = 0.0258
-        self.fr_n3ish = 0.02581
-
-        self.fr_N = 0
-        self.fr_N_up = 0
-        self.N_up = 0
-        self.act_N_up_each_layer = []
-        self.N_actual_up = 0
-
-        # ===================================================================
-        ''' Phosphorus Uptake Data '''
-
-        self.beta_p = 10  # corn
-
-        self.bio_P_opt = 0
-        self.bio_P = 0
-
-        self.fr_PHU_50 = 0.5
-        self.fr_PHU_100 = 1.0
-        self.fr_p1 = 0.0074
-        self.fr_p2 = 0.0037
-        self.fr_p3 = 0.0035
-        self.fr_p3ish = 0.00351
-
-        self.fr_P = 0
-        self.P_up = 0
-        self.act_P_up_each_layer = []
-        self.P_act_up = 0
-
-        # ===================================================================
-        ''' Yields Data '''
-
-        self.HI_max = 0
-        self.HI_min = 0.01
-        self.HI_actual = 0
-        self.HI_opt = 0.31
-
-        self.harvest_eff = 0.9
-
-        self.gamma_wu = 0
-
-        self.bio_AG = 0
-        self.yield_max = 0
-        self.yield_actual = 0
-        self.yield_N = 0
-        self.yield_P = 0
-
-        self.yield_annual = 0
-
-
-#
-# Crop object populated with Alfalfa data
-#
-class Alfalfa:
-
-    def __init__(self, data):
-        """GENERAL PLANT INFO"""
-
-        alfalfa_data = data['alfalfa']
-        self.grow_years = alfalfa_data['grow_years']
-        self.repeat = alfalfa_data['repeat']
-        self.planting_date = alfalfa_data['planting_date']
-        self.harvest_date = alfalfa_data['harvest_date']
-
-        if alfalfa_data['harvest_type'] != 'optimal':
-            print('Perennial crops are always optimally harvested')
-
-        self.harvest_type = 'optimal'
-
-        self.crop_name = 'alfalfa'
-        self.crop_type = 'perennial'
-
-        self.kill_day = -1
-        self.kill_year = False
-        self.planted = False
-        self.growing = False
-
-        self.fix_nitrogen = True
-        # ===================================================================
-        ''' HEAT UNIT DATA '''
-
-        # Inputs
-        self.T_base_min = 4
-        self.T_base_max = 32  # until dormancy
-        self.PHU = 800  # still unknown
-
-        # Internally calculated inputs
-        self.accumulated_HU = 0.0
-        self.prev_accumulated_HU = 0.0
-
-        # Outputs
-        self.fr_PHU = 0.0
-        self.prev_fr_PHU = 0.0
-
-        # ===================================================================
-        ''' LEAF AREA INDEX (LAI) DATA '''
-
-        # Inputs
-        self.fr_PHU_1 = 0.15
-        self.fr_PHU_2 = 0.50
-        self.fr_LAI_1 = 0.01
-        self.fr_LAI_2 = 0.95
-        self.fr_PHU_sen = 0.90
-        self.fr_PHU_harvest = 1.0  # TODO: If alfalfa is a hay, PHU to cut is 0.6
-        self.LAI_max = 4
-        self.LAI_min = 0.75
-
-        # Internally calculated inputs
-        self.prev_fr_LAI_max = 0
-        self.fr_LAI_max = 0
-
-        # Outputs
-        self.prev_LAI_actual = 0
-        self.LAI_actual = 0
-
-        # ===================================================================
-        ''' ROOT DEPTH DATA '''
-
-        # Inputs
-        self.z_root_max = 3000  # maximum depth of root development
-
-        # Internally calculated inputs
-        self.fr_root = 0
-
-        # Outputs
-        self.z_root = 0
-
-        # ===================================================================
-        ''' BIOMASS DATA '''
-
-        # Inputs
-        self.kl = 0.65
-        self.RUE = 20
-        self.T_opt = 25
-
-        # Internally calculated inputs
-        self.gamma_reg = 0
-        self.d_biomass_max = 0
-        self.d_biomass_actual = 0.0
-
-        # Outputs
-        self.biomass_actual = 0
-        self.prev_biomass_actual = 0
-
-        # ===================================================================
-        ''' Soil Water Uptake Data '''
-
-        self.beta_w = 10  # water-use distribution parameter  # corn
-        self.epco = 0.5  # corn
-
-        self.water_actual_up = 0
-        self.water_uptake_each_layer = []
-
-        # ===================================================================
-        ''' Nitrogen Uptake Data '''
-
-        self.beta_n = 10  # corn
-
-        self.bio_N_opt = 0
-        self.bio_N = 0
-
-        self.fr_n1 = 0.0417
-        self.fr_n2 = 0.0290
-        self.fr_n3 = 0.0200
-        self.fr_n3ish = 0.02001
-
-        self.fr_N = 0
-        self.fr_N_up = 0
-        self.N_up = 0
-        self.act_N_up_each_layer = []
-        self.N_actual_up = 0
-
-        # ===================================================================
-        ''' Phosphorus Uptake Data '''
-
-        self.beta_p = 10  # corn
-
-        self.bio_P_opt = 0
-        self.bio_P = 0
-
-        self.fr_PHU_50 = 0.5
-        self.fr_PHU_100 = 1.0
-        self.fr_p1 = 0.0035
-        self.fr_p2 = 0.0028
-        self.fr_p3 = 0.0020
-        self.fr_p3ish = 0.00201
-
-        self.fr_P = 0
-        self.P_up = 0
-        self.act_P_up_each_layer = []
-        self.P_act_up = 0
-
-        # ===================================================================
-        ''' Yields Data '''
-
-        self.HI_max = 0
-        self.HI_min = 0.9
-        self.HI_actual = 0
-        self.HI_opt = 0.9
-
-        self.harvest_eff = 0.9
 
         self.gamma_wu = 0
